adding more compiler tests
