+++ src/state/reducers/alignments/verse.js
@@ -24,6 +24,7 @@ import renderedAlignmentsReducer, * as fromRenderedAlignments
 import renderedAlignmentReducer from './renderedAlignment';
 import _ from 'lodash';
 import suggestionReducer from './suggestion';
+import compile from './compile';
 
 /**
  * Compares two alignments for sorting
@@ -455,225 +456,226 @@ export const compileLegacy = (rendered, alignments) => {
   };
 };
 
-export const compile = (renders, alignments) => {
-  // const compiledIndices = {};
-  const processedAlignments = [];
-  let approvedAlignments = [];
-  const compiledRenders = {};
-
-  // index the alignment keys
-  // TODO: move this into the loop below.
-  const alignmentIndexMap = {};
-  for (let i = 0; i < alignments.length; i++) {
-    alignmentIndexMap[alignments[i].sourceNgram.join('|')] = i;
-  }
-
-  // index renders by alignment
-  const siblingIndex = {};
-  for (let rIndex = 0; rIndex < renders.length; rIndex++) {
-    const r = renders[rIndex];
-    for (const aIndex of r.alignments) {
-      if (!siblingIndex[aIndex]) {
-        siblingIndex[aIndex] = [];
-      }
-      siblingIndex[aIndex].push(rIndex);
-    }
-  }
-
-  for (let rIndex = 0; rIndex < renders.length; rIndex++) {
-    const r = renders[rIndex];
-    const isSuggestion = r.suggestion !== undefined;
-    // if (!compiledIndices[rIndex]) {
-    //   compiledIndices[rIndex] = [];
-    // }
-    if (rIndex in compiledRenders) {
-      continue;
-    }
-
-    if (!isSuggestion) {
-      // compile approved alignments
-      // TRICKY: approved suggestions only have a single alignment
-      approvedAlignments = approvedAlignments.concat(
-        siblingIndex[r.alignments[0]]);
-      compileApprovedRender(rIndex, renders, siblingIndex, compiledRenders);
-    } else {
-      // compile everything else
-      for (const aIndex of r.alignments) {
-        // checks if a sibling of this render has been approved.
-        const isSiblingApproved = approvedAlignments.indexOf(aIndex) >= 0;
-        const alreadyCompiled = processedAlignments.indexOf(aIndex) >= 0;
-
-        // update index mapping
-        if (alreadyCompiled && !isSiblingApproved) {
-          // compiledIndices[rIndex].push(processedAlignments.indexOf(aIndex));
-          // TRICKY: suggested alignment splits will cause the alignment to appear multiple times
-          // TODO: we need to add this to the listed renderedIndex of the alignment
-          continue;
-        } else {
-          // compiledIndices[rIndex].push(processedAlignments.length);
-          // processedAlignments.push(aIndex);
-        }
-
-        if (!isSiblingApproved) {
-          // compile un-approved suggestions
-          const alignment = alignments[aIndex];
-          // TRICKY: as an un-approved suggestion there may be multiple alignments
-          let values = [];
-          if (compiledRenders[rIndex]) {
-            values = compiledRenders[rIndex].values;
-          }
-
-          compiledRenders[rIndex] = {
-            isSuggestion: true,
-            index: rIndex,
-            values: [
-              ...values, {
-                ..._.cloneDeep(alignment),
-                renderedIndex: rIndex
-              }]
-          };
-        } else {
-          // compile partially approved suggestions (splits)
-          compiledRenders[rIndex] = {
-            isSuggestion: false,
-            index: rIndex,
-            values: {
-              renderedIndex: rIndex,
-              sourceNgram: [...r.sourceNgram],
-              targetNgram: []
-            }
-          };
-        }
-      }
-    }
-  }
-
-  const compiledIndices = {};
-  const compiledAlignments = [];
-  const flattenedAlignments = _.flatten(
-    _.sortBy(Object.values(compiledRenders), [o => o.index]).
-      map(o => o.values));
-
-  // remove duplicates
-
-  // generate index map
-  const addedAlignments = [];
-  for (let i = 0; i < flattenedAlignments.length; i++) {
-    const a = {...flattenedAlignments[i]};
-    const id = a.sourceNgram.join('|');
-    const originalIndex = alignmentIndexMap[id];
-
-    if (!compiledIndices[a.renderedIndex]) {
-      compiledIndices[a.renderedIndex] = [];
-    }
-
-    if (addedAlignments.indexOf(id) >= 0) {
-      compiledIndices[a.renderedIndex].push(addedAlignments.indexOf(id));
-      continue;
-    }
-
-    if (originalIndex && siblingIndex[originalIndex]) {
-      // map old indices
-      for (const renderedIndex of siblingIndex[originalIndex]) {
-        if (!compiledIndices[renderedIndex]) {
-          compiledIndices[renderedIndex] = [];
-        }
-        compiledIndices[renderedIndex].push(i);
-      }
-    } else {
-      // TODO: only put a rendered index on the ones we need.
-      // map new indices
-      compiledIndices[a.renderedIndex].push(i);
-    }
-    delete a.renderedIndex;
-    // if(addedAlignments.indexOf(id) === -1) {
-    compiledAlignments.push(a);
-    addedAlignments.push(id);
-    // } else {
-    //   compiledIndices[addedAlignments.indexOf(id)].push(i);
-    // }
-  }
-
-  // const compiledAlignments = flattenedAlignments.map((a, i) => {
-  //   if (!compiledIndices[a.renderedIndex]) {
-  //     compiledIndices[a.renderedIndex] = [];
-  //   }
-  //   compiledIndices[a.renderedIndex].push(i);
-  //   delete a.renderedIndex;
-  //   return a;
-  // });
-
-  return {
-    alignments: compiledAlignments,
-    indices: compiledIndices
-  };
-};
-
-/**
- * Recursively compiles an approved rendered alignment
- * @param rIndex
- * @param renders
- * @param siblingIndex
- * @param compiledRenders
- */
-const compileApprovedRender = (
-  rIndex, renders, siblingIndex, compiledRenders) => {
-  const r = renders[rIndex];
-  compiledRenders[rIndex] = {
-    isSuggestion: false,
-    index: rIndex,
-    values: [
-      {
-        renderedIndex: rIndex,
-        sourceNgram: [...r.sourceNgram],
-        targetNgram: [...r.targetNgram]
-      }]
-  };
-  for (const aIndex of r.alignments) {
-    for (const sIndex of siblingIndex[aIndex]) {
-      if (sIndex !== rIndex) {
-        compileSplitSiblings(sIndex, renders, siblingIndex, compiledRenders);
-      }
-    }
-  }
-};
-
-/**
- * Compiles the siblings of a split alignment suggestion
- * @param rIndex
- * @param renders
- * @param siblingIndex
- * @param compiledRenders
- */
-const compileSplitSiblings = (
-  rIndex, renders, siblingIndex, compiledRenders) => {
-  const r = renders[rIndex];
-
-  // only compile un-approved suggestions
-  if ((rIndex in compiledRenders && !compiledRenders[rIndex].isSuggestion)
-    || (r.suggestion === undefined)) {
-    return;
-  }
-
-  compiledRenders[rIndex] = {
-    isSuggestion: false,
-    index: rIndex,
-    values: [
-      {
-        renderedIndex: rIndex,
-        sourceNgram: [...r.sourceNgram],
-        targetNgram: []
-      }]
-  };
-
-  for (const aIndex of r.alignments) {
-    for (const sIndex of siblingIndex[aIndex]) {
-      if (sIndex !== rIndex) {
-        compileSplitSiblings(sIndex, renders, siblingIndex, compiledRenders);
-      }
-    }
-  }
-};
+// export const compile = (renders, alignments) => {
+//   // const compiledIndices = {};
+//   // const processedAlignments = [];
+//   let approvedAlignments = [];
+//   const compiledRenders = {};
+//
+//   // index the alignment keys
+//   // TODO: move this into the loop below.
+//   const alignmentIndexMap = {};
+//   for (let i = 0; i < alignments.length; i++) {
+//     alignmentIndexMap[alignments[i].sourceNgram.join('|')] = i;
+//   }
+//
+//   // index renders by alignment
+//   const siblingIndex = {};
+//   for (let siblingPos = 0; siblingPos < renders.length; siblingPos++) {
+//     for (const aIndex of renders[siblingPos].alignments) {
+//       if (!siblingIndex[aIndex]) {
+//         siblingIndex[aIndex] = [];
+//       }
+//       siblingIndex[aIndex].push(siblingPos);
+//     }
+//   }
+//
+//
+//   for (let renderPos = 0; renderPos < renders.length; renderPos++) {
+//     const r = renders[renderPos];
+//     const isSuggestion = r.suggestion !== undefined;
+//     // if (!compiledIndices[rIndex]) {
+//     //   compiledIndices[rIndex] = [];
+//     // }
+//     if (renderPos in compiledRenders) {
+//       continue;
+//     }
+//
+//     if (!isSuggestion) {
+//       // compile approved alignments
+//       // TRICKY: approved suggestions only have a single alignment
+//       approvedAlignments = approvedAlignments.concat(
+//         siblingIndex[r.alignments[0]]);
+//       compileApprovedRender(renderPos, renders, siblingIndex, compiledRenders);
+//     } else {
+//       // compile everything else
+//       for (const aIndex of r.alignments) {
+//         // checks if a sibling of this render has been approved.
+//         const isSiblingApproved = approvedAlignments.indexOf(aIndex) >= 0;
+//         const alignment = alignments[aIndex];
+//         const suggestedTaretTokens = _.difference(alignment.targetNgram,
+//           r.targetNgram);
+//         const isAlignmentUpdated = _.difference(suggestedTaretTokens,
+//           r.suggestedTargetTokens).length > 0;
+//         // const alreadyCompiled = processedAlignments.indexOf(aIndex) >= 0;
+//
+//         // update index mapping
+//         // if (alreadyCompiled && !isSiblingApproved) {
+//         // compiledIndices[rIndex].push(processedAlignments.indexOf(aIndex));
+//         // TRICKY: suggested alignment splits will cause the alignment to appear multiple times
+//         // TODO: we need to add this render index to the listed renderedIndex of the alignment
+//         // continue;
+//         // } else {
+//         // compiledIndices[rIndex].push(processedAlignments.length);
+//         // processedAlignments.push(aIndex);
+//         // }
+//
+//         if(isAlignmentUpdated) {
+//           // compile partially approved suggestions (splits)
+//           compiledRenders[renderPos] = {
+//             isSuggestion: false,
+//             index: renderPos,
+//             values: {
+//               renderedIndex: renderPos,
+//               sourceNgram: [...r.sourceNgram],
+//               targetNgram: [...suggestedTaretTokens]
+//             }
+//           };
+//         } else if (isSiblingApproved) {
+//           // compile partially approved suggestions (splits)
+//           compiledRenders[renderPos] = {
+//             isSuggestion: false,
+//             index: renderPos,
+//             values: {
+//               renderedIndex: renderPos,
+//               sourceNgram: [...r.sourceNgram],
+//               targetNgram: []
+//             }
+//           };
+//         } else {
+//           // compile un-approved suggestions
+//           // TRICKY: as an un-approved suggestion there may be multiple alignments
+//           // so we must append the values
+//           let values = [];
+//           if (compiledRenders[renderPos]) {
+//             values = compiledRenders[renderPos].values;
+//           }
+//
+//           compiledRenders[renderPos] = {
+//             isSuggestion: true,
+//             index: renderPos,
+//             values: [
+//               ...values, {
+//                 ..._.cloneDeep(alignment),
+//                 renderedIndex: renderPos
+//               }]
+//           };
+//         }
+//       }
+//     }
+//   }
+//
+//   const compiledIndices = {};
+//   const compiledAlignments = [];
+//   const flattenedAlignments = _.flatten(
+//     _.sortBy(Object.values(compiledRenders), [o => o.index]).
+//       map(o => o.values));
+//
+//   // generate index map
+//   const addedAlignments = [];
+//   for (let i = 0; i < flattenedAlignments.length; i++) {
+//     const a = {...flattenedAlignments[i]};
+//     const id = a.sourceNgram.join('|');
+//     const originalIndex = alignmentIndexMap[id];
+//
+//     if (!compiledIndices[a.renderedIndex]) {
+//       compiledIndices[a.renderedIndex] = [];
+//     }
+//
+//     // skip duplicates
+//     if (addedAlignments.indexOf(id) >= 0) {
+//       compiledIndices[a.renderedIndex].push(addedAlignments.indexOf(id));
+//       continue;
+//     }
+//
+//     if (originalIndex && siblingIndex[originalIndex]) {
+//       // map old indices
+//       for (const renderedIndex of siblingIndex[originalIndex]) {
+//         if (!compiledIndices[renderedIndex]) {
+//           compiledIndices[renderedIndex] = [];
+//         }
+//         compiledIndices[renderedIndex].push(i);
+//       }
+//     } else {
+//       // map new indices
+//       compiledIndices[a.renderedIndex].push(i);
+//     }
+//     delete a.renderedIndex;
+//     compiledAlignments.push(a);
+//     addedAlignments.push(id);
+//   }
+//
+//   return {
+//     alignments: compiledAlignments,
+//     indices: compiledIndices
+//   };
+// };
+//
+// /**
+//  * Recursively compiles an approved rendered alignment
+//  * @param rIndex
+//  * @param renders
+//  * @param siblingIndex
+//  * @param compiledRenders
+//  */
+// const compileApprovedRender = (
+//   rIndex, renders, siblingIndex, compiledRenders) => {
+//   const r = renders[rIndex];
+//   compiledRenders[rIndex] = {
+//     isSuggestion: false,
+//     index: rIndex,
+//     values: [
+//       {
+//         renderedIndex: rIndex,
+//         sourceNgram: [...r.sourceNgram],
+//         targetNgram: [...r.targetNgram]
+//       }]
+//   };
+//   for (const aIndex of r.alignments) {
+//     for (const sIndex of siblingIndex[aIndex]) {
+//       if (sIndex !== rIndex) {
+//         compileSplitSiblings(sIndex, renders, siblingIndex, compiledRenders);
+//       }
+//     }
+//   }
+// };
+//
+// /**
+//  * Compiles the siblings of a split alignment suggestion
+//  * @param rIndex
+//  * @param renders
+//  * @param siblingIndex
+//  * @param compiledRenders
+//  */
+// const compileSplitSiblings = (
+//   rIndex, renders, siblingIndex, compiledRenders) => {
+//   const r = renders[rIndex];
+//
+//   // only compile un-approved suggestions
+//   if ((rIndex in compiledRenders && !compiledRenders[rIndex].isSuggestion)
+//     || (r.suggestion === undefined)) {
+//     return;
+//   }
+//
+//   compiledRenders[rIndex] = {
+//     isSuggestion: false,
+//     index: rIndex,
+//     values: [
+//       {
+//         renderedIndex: rIndex,
+//         sourceNgram: [...r.sourceNgram],
+//         targetNgram: []
+//       }]
+//   };
+//
+//   for (const aIndex of r.alignments) {
+//     for (const sIndex of siblingIndex[aIndex]) {
+//       if (sIndex !== rIndex) {
+//         compileSplitSiblings(sIndex, renders, siblingIndex, compiledRenders);
+//       }
+//     }
+//   }
+// };
 
 /**
  * Returns the source tokens being aligned.
